+++ src/ql/arch/cc/pass/gen/vq1asm/detail/codegen.cc
@@ -658,9 +658,9 @@ void Codegen::forStart(const Str &label, UInt iterations) {
     comment(QL_SS2S("# FOR_START(" << iterations << ")"));
     // FIXME: reserve register
     emit("", "move", QL_SS2S(iterations << ",R62"), "# R62 is the 'for loop counter'");        // FIXME: fixed reg, no nested for loops (not supported by program.cc either)
-    emit((label+":"), "", "", "# ");        // just a label
+    emit((label+":"), "", "", "# ");
 #if OPT_PRAGMA
-    pragmaForLabel = label;        // remind label for pragma/break FIXME: implement properly later on
+    pragmaLoopLabel.push_back(label);        // remind label for pragma/break FIXME: implement properly later on
 #endif
 }
 
@@ -669,19 +669,27 @@ void Codegen::forEnd(const Str &label) {
     // FIXME: free register
     emit("", "loop", QL_SS2S("R62,@" << label), "# R62 is the 'for loop counter'");        // FIXME: fixed reg, no nested for loops (not supported by program.cc either)
 #if OPT_PRAGMA
-    emit((label+"_end:"), "", "", "# ");                          // NB: just a label
+    emit((label+"_end:"), "", "", "# ");    // label for 'break'
+    pragmaLoopLabel.pop_back();
 #endif
 }
 
 void Codegen::doWhileStart(const Str &label) {
     comment("# DO_WHILE_START");
-    emit((label+":"), "", "", "# ");                              // NB: just a label
+    emit((label+":"), "", "", "# ");
+#if OPT_PRAGMA
+    pragmaLoopLabel.push_back(label);        // remind label for pragma/break FIXME: implement properly later on
+#endif
 }
 
 void Codegen::doWhileEnd(const Str &label, UInt op0, const Str &opName, UInt op1) {
     comment(QL_SS2S("# DO_WHILE_END(R" << op0 << " " << opName << " R" << op1 << ")"));
     emit("", "jmp", QL_SS2S("@" << label), "# FIXME: we don't support conditions, just an endless loop'");        // FIXME: just endless loop
     QL_WOUT("CC backend ignores condition of do while loop");
+#if OPT_PRAGMA
+    emit((label+"_end:"), "", "", "# ");    // label for 'break'
+    pragmaLoopLabel.pop_back();
+#endif
 }
 
 void Codegen::comment(const Str &c) {
@@ -749,7 +757,7 @@ void Codegen::emitProgramStart(const Str &progName) {
     // NB: new seq_bar semantics (firmware from 20191219 onwards)
     comment("# synchronous start and latency compensation");
     emit("",                "seq_bar",  "",                 "# synchronization, delay set externally through SET_SEQ_BAR_CNT");
-
+    emit("",                "seq_out",  "0x00000000,1",     "# allows monitoring actual start time using trace unit");
     emit("__mainLoop:",     "",         "",                 "# ");    // FIXME: __mainLoop should be a forbidden kernel name
 
 #if OPT_FEEDBACK
@@ -881,7 +889,7 @@ void Codegen::emitPragma(
     Int pragmaBreakVal = json_get<Int>(pragma, "break", "pragma of unknown instruction");        // FIXME: we don't know which instruction we're dealing with, so better move
     UInt smAddr = pragmaSmBit / 32;    // 'seq_cl_sm' is addressable in 32 bit words
     UInt mask = 1ul << (pragmaSmBit % 32);
-    std::string label = pragmaForLabel+"_end";        // FIXME: must match label set in forEnd(), assumes we are actually inside a for loop
+    std::string label = pragmaLoopLabel.back() + "_end";        // FIXME: must match label set in forEnd(), assumes we are actually inside a for loop
 
     // emit code for pragma "break". NB: code is identical for all instruments
     // FIXME: verify that instruction duration matches actual time